changed track scales to 8 + some odometry

+++ sdsim/Assets/Scripts/SocketIO/JSONObject/VectorTemplates.cs
@@ -217,9 +217,9 @@ public static partial class JSONTemplates {
 	}
 	
 	public static Keyframe ToKeyframe(JSONObject obj){
-		Keyframe k = new Keyframe(obj.HasField("time")? obj.GetField("time").n : 0, obj.HasField("value")? obj.GetField("value").n : 0);
-		if(obj.HasField("inTangent")) k.inTangent = obj.GetField("inTangent").n;
-		if(obj.HasField("outTangent")) k.outTangent = obj.GetField("outTangent").n;
+		Keyframe k = new Keyframe((float)(obj.HasField("time")? obj.GetField("time").n : 0), (float)(obj.HasField("value")? obj.GetField("value").n : 0));
+		if(obj.HasField("inTangent")) k.inTangent = (float)obj.GetField("inTangent").n;
+		if(obj.HasField("outTangent")) k.outTangent = (float)obj.GetField("outTangent").n;
 		//if(obj.HasField("tangentMode")) k.tangentMode = (int)obj.GetField("tangentMode").n;
 		
 		return k;
